Split ld lowering into form selection and encoding
